use read_key
