+++ crates/sui-bridge/Cargo.toml
@@ -25,6 +25,7 @@ rocksdb.workspace = true
 typed-store.workspace = true
 mysten-metrics.workspace = true
 sui-sdk.workspace = true
+sui-keys.workspace = true
 sui-config.workspace = true
 clap.workspace = true
 tracing.workspace = true
