+++ crates/sui-bridge/src/config.rs
@@ -10,8 +10,6 @@ use crate::types::{is_route_valid, BridgeAction};
 use anyhow::anyhow;
 use ethers::providers::Middleware;
 use ethers::types::Address as EthAddress;
-use fastcrypto::encoding::{Encoding, Hex};
-use fastcrypto::traits::{EncodeDecodeBase64, KeyPair};
 use futures::{future, StreamExt};
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
@@ -21,12 +19,14 @@ use std::str::FromStr;
 use std::sync::Arc;
 use sui_config::Config;
 use sui_json_rpc_types::Coin;
+use sui_keys::keypair_file::read_key;
 use sui_sdk::apis::CoinReadApi;
 use sui_sdk::{SuiClient as SuiSdkClient, SuiClientBuilder};
 use sui_types::base_types::ObjectRef;
 use sui_types::base_types::{ObjectID, SuiAddress};
 use sui_types::bridge::BridgeChainId;
-use sui_types::crypto::{SuiKeyPair, ToFromBytes};
+use sui_types::crypto::KeypairTraits;
+use sui_types::crypto::SuiKeyPair;
 use sui_types::digests::{get_mainnet_chain_identifier, get_testnet_chain_identifier};
 use sui_types::event::EventID;
 use sui_types::object::Owner;
@@ -67,10 +67,10 @@ pub struct SuiConfig {
     pub sui_rpc_url: String,
     /// The expected BridgeChainId on Sui side.
     pub sui_bridge_chain_id: u8,
-    /// Path of the file where bridge client key (any SuiKeyPair) is stored as Base64 encoded `flag || privkey`.
-    /// If `run_client` is true, and this is None, then use `bridge_authority_key_path_base64_raw` as client key.
+    /// Path of the file where bridge client key (any SuiKeyPair) is stored.
+    /// If `run_client` is true, and this is None, then use `bridge_authority_key_path` as client key.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub bridge_client_key_path_base64_sui_key: Option<PathBuf>,
+    pub bridge_client_key_path: Option<PathBuf>,
     /// The gas object to use for paying for gas fees for the client. It needs to
     /// be owned by the address associated with bridge client key. If not set
     /// and `run_client` is true, it will query and use the gas object with highest
@@ -96,9 +96,9 @@ pub struct BridgeNodeConfig {
     pub server_listen_port: u16,
     /// The port that for metrics server.
     pub metrics_port: u16,
-    /// Path of the file where bridge authority key (Secp256k1) is stored as Base64 encoded `privkey`.
-    pub bridge_authority_key_path_base64_raw: PathBuf,
-    /// Whether to run client. If true, `bridge_client_key_path_base64_sui_key`,
+    /// Path of the file where bridge authority key (Secp256k1) is stored.
+    pub bridge_authority_key_path: PathBuf,
+    /// Whether to run client. If true, `sui.bridge_client_key_path`
     /// and `db_path` needs to be provided.
     pub run_client: bool,
     /// Path of the client storage. Required when `run_client` is true.
@@ -129,8 +129,10 @@ impl BridgeNodeConfig {
             ));
         };
 
-        let bridge_authority_key =
-            read_bridge_authority_key(&self.bridge_authority_key_path_base64_raw)?;
+        let bridge_authority_key = match read_key(&self.bridge_authority_key_path, true)? {
+            SuiKeyPair::Secp256k1(key) => key,
+            _ => unreachable!("we required secp256k1 key in `read_key`"),
+        };
 
         // we do this check here instead of `prepare_for_sui` below because
         // that is only called when `run_client` is true.
@@ -292,13 +294,9 @@ impl BridgeNodeConfig {
         &self,
         sui_client: Arc<SuiClient<SuiSdkClient>>,
     ) -> anyhow::Result<(SuiKeyPair, SuiAddress, ObjectRef)> {
-        let bridge_client_key = match &self.sui.bridge_client_key_path_base64_sui_key {
-            None => {
-                let bridge_client_key =
-                    read_bridge_authority_key(&self.bridge_authority_key_path_base64_raw)?;
-                Ok(SuiKeyPair::from(bridge_client_key))
-            }
-            Some(path) => read_bridge_client_key(path),
+        let bridge_client_key = match &self.sui.bridge_client_key_path {
+            None => read_key(&self.bridge_authority_key_path, true),
+            Some(path) => read_key(path, false),
         }?;
 
         // If bridge chain id is Sui Mainent or Testnet, we expect to see chain
@@ -386,81 +384,6 @@ pub struct BridgeClientConfig {
     pub eth_contracts_start_block_fallback: u64,
     pub eth_contracts_start_block_override: Option<u64>,
     pub sui_bridge_module_last_processed_event_id_override: Option<EventID>,
-}
-
-/// Read Bridge Authority key (Secp256k1KeyPair) from a file.
-/// BridgeAuthority key is stored as base64 encoded `privkey`.
-pub fn read_bridge_authority_key(path: &PathBuf) -> Result<BridgeAuthorityKeyPair, anyhow::Error> {
-    if !path.exists() {
-        return Err(anyhow::anyhow!(
-            "Bridge authority key file not found at path: {:?}",
-            path
-        ));
-    }
-    let contents = std::fs::read_to_string(path)?;
-
-    BridgeAuthorityKeyPair::decode_base64(contents.as_str().trim())
-        .map_err(|e| anyhow!("Error decoding authority key: {:?}", e))
-}
-
-/// Read Bridge client key (any SuiKeyPair) from a file.
-/// Read from file as Base64 encoded `flag || privkey`.
-pub fn read_bridge_client_key(path: &PathBuf) -> Result<SuiKeyPair, anyhow::Error> {
-    if !path.exists() {
-        return Err(anyhow::anyhow!(
-            "Bridge client key file not found at path: {:?}",
-            path
-        ));
-    }
-    let contents = std::fs::read_to_string(path)?;
-
-    SuiKeyPair::decode_base64(contents.as_str().trim())
-        .map_err(|e| anyhow!("Error decoding authority key: {:?}", e))
-}
-
-/// Read a SuiKeyPair from a file. The content could be any of the following:
-/// - Base64 encoded `flag || privkey` for ECDSA key
-/// - Base64 encoded `privkey` for Raw key
-/// - Bech32 encoded private key prefixed with `suiprivkey`
-/// - Hex encoded `privkey` for Raw key
-/// If `require_secp256k1` is true, it will return an error if the key is not Secp256k1.
-pub fn read_key(path: &PathBuf, require_secp256k1: bool) -> Result<SuiKeyPair, anyhow::Error> {
-    if !path.exists() {
-        return Err(anyhow::anyhow!("Key file not found at path: {:?}", path));
-    }
-    let file_contents = std::fs::read_to_string(path)?;
-    let contents = file_contents.as_str().trim();
-
-    // Try base64 encoded SuiKeyPair `flag || privkey`
-    if let Ok(key) = SuiKeyPair::decode_base64(contents) {
-        if require_secp256k1 && !matches!(key, SuiKeyPair::Secp256k1(_)) {
-            return Err(anyhow!("Key is not Secp256k1"));
-        }
-        return Ok(key);
-    }
-
-    // Try base64 encoded Raw Secp256k1 key `privkey`
-    if let Ok(key) = BridgeAuthorityKeyPair::decode_base64(contents) {
-        return Ok(SuiKeyPair::Secp256k1(key));
-    }
-
-    // Try Bech32 encoded 33-byte `flag || private key` starting with `suiprivkey`A prefix.
-    // This is the format of a private key exported from Sui Wallet or sui.keystore.
-    if let Ok(key) = SuiKeyPair::decode(contents) {
-        if require_secp256k1 && !matches!(key, SuiKeyPair::Secp256k1(_)) {
-            return Err(anyhow!("Key is not Secp256k1"));
-        }
-        return Ok(key);
-    }
-
-    // Try hex encoded Raw key `privkey`
-    if let Ok(bytes) = Hex::decode(contents).map_err(|e| anyhow!("Error decoding hex: {:?}", e)) {
-        if let Ok(key) = BridgeAuthorityKeyPair::from_bytes(&bytes) {
-            return Ok(SuiKeyPair::Secp256k1(key));
-        }
-    }
-
-    Err(anyhow!("Error decoding key from {:?}", path))
 }
 
 #[serde_as]
